+++ src/main/java/persistence/study/reflection/PrintView.java
@@ -1,4 +1,4 @@
-package persistence.study;
+package persistence.study.reflection;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
