+++ src/test/java/persistence/study/reflection/ReflectionTest.java
@@ -1,4 +1,4 @@
-package persistence.study;
+package persistence.study.reflection;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
